+++ templates/utils/scripts/global.d.ts
@@ -1,0 +1,4 @@
+/** Object storing module functions */
+declare var modules: { [index:string]: Function };
+
+declare module 'uuid/v4';