+++ CHANGELOG.md
@@ -3,15 +3,20 @@
 ### Adds
 
 - Adds: `pjax.twig` layout that requires `craft.app.request.isPjax`
+- Adds: `/templates/utils` directory
+- Adds: `/templates/lib` directory for storing Objects, Components, and Globals
 
 ### Fixes
 
 - Fixes: `ajax.twig` layout requires `craft.app.request.isAjax`
 - Fixes: updates the base 404 page
 - Fixes: cleans up the base homepage
+- Fixes: webpack config
+- Fixes: typescript config
 
 ### Removes
 
 - Removes: base complex content directory
 - Removes: initial form SVGs
 - Removes: base form macros
+- Removes: app directory